requirements/base.txt
Django==5.1.4
django-ninja==1.1.0
